third_party/WebKit/LayoutTests/fast/events/touch/gesture/gesture-scrollbar-textarea-expected.txt

This tests scroll gesture events on a textarea scrollbar. The textarea below should be slightly scrolled down if successful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS textArea.scrollTop is 0
PASS textArea.scrollTop is >= 1
PASS successfullyParsed is true

TEST COMPLETE
